requirements.txt
# Core Telegram Bot dependencies
python-telegram-bot>=20.7
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.2.0

# Text Recognition (OCR)
pytesseract>=0.3.10

# Audio Processing
numpy>=1.24.0
scipy>=1.11.0

# Text-to-Speech
gtts>=2.5.0

# Language Processing & Grammar
langdetect>=1.0.9
pyspellchecker>=0.6.3
spacy>=3.6.0

# AI/ML Dependencies (CPU optimized for ARM/x64)
torch>=2.1.0
torchaudio>=2.1.0
openai-whisper>=20231117

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development dependencies (опционально)
black>=23.0.0      # Форматирование кода
flake8>=6.0.0      # Линтинг
mypy>=1.7.0        # Статическая типизация

# HTTP client
httpx>=0.25.0       # Асинхронный HTTP клиент

# File handling
aiofiles>=23.0.0    # Асинхронная работа с файлами

# Configuration management
pydantic>=2.5.0     # Валидация конфигурации
pydantic-settings>=2.1.0  # Настройки из .env файлов

# Utilities
python-dateutil>=2.8.0  # Работа с датами

# Kafka
kafka-python>=2.0.2

# Stable Diffusion XL
diffusers>=0.25.0
transformers>=0.25.0
accelerate>=0.25.0
